Fix file moves and add file comments
