+++ apps/database/src/scripts/reg-fetchers.ts
@@ -1,4 +1,9 @@
-// Functions to fetch course data from the registrar API
+/**
+ * @file scripts/reg-fetchers.ts
+ * @description Functions for fetching course listings and details from
+ * the registrar API.
+ * @author Joshua Lau
+ */
 
 import type {
     RegCourseDetails,
